requirements.txt
# AI・機械学習関連
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
openai>=1.0.0

# データベース関連
pymongo>=4.6.0
motor>=3.3.0

# Web API関連
fastapi>=0.104.0
uvicorn>=0.24.0
line-bot-sdk>=3.9.0
requests>=2.31.0

# 環境設定・セキュリティ
python-dotenv>=1.0.0
google-cloud-secret-manager>=2.18.0
google-cloud-functions>=1.15.0

# データ処理関連
pandas>=2.1.0
numpy>=1.25.0
pydantic>=2.5.0

# 開発・テスト用
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0